feat(v2/appconfigdata): fixing small things
