+++ tests/e2e/parameters/test_appconfig.py
@@ -1,0 +1,36 @@
+import json
+
+import pytest
+
+from tests.e2e.utils import data_fetcher
+
+
+@pytest.fixture
+def parameter_appconfig_freeform_handler_fn_arn(infrastructure: dict) -> str:
+    return infrastructure.get("ParameterAppconfigFreeformHandlerArn", "")
+
+
+@pytest.fixture
+def parameter_appconfig_freeform_handler_fn(infrastructure: dict) -> str:
+    return infrastructure.get("ParameterAppconfigFreeformHandler", "")
+
+
+@pytest.fixture
+def parameter_appconfig_freeform_value(infrastructure: dict) -> str:
+    return infrastructure.get("AppConfigConfigurationValue", "")
+
+
+def test_get_parameter_appconfig_freeform(
+    parameter_appconfig_freeform_handler_fn_arn: str, parameter_appconfig_freeform_value: str
+):
+    # GIVEN
+    payload = json.dumps({"name": "profilee2e", "environment": "enve2e", "application": "appe2e"})
+    expected_return = parameter_appconfig_freeform_value
+
+    # WHEN
+    parameter_execution, _ = data_fetcher.get_lambda_response(
+        lambda_arn=parameter_appconfig_freeform_handler_fn_arn, payload=payload
+    )
+    parameter_value = parameter_execution["Payload"].read().decode("utf-8")
+
+    assert parameter_value == expected_return
